delete bei produkt teiweise fertig toString bei Stock ueberschreiben

+++ src/main/java/de/pdbm/starter/business/messages/service/ProductService.java
@@ -1,12 +1,16 @@
 package de.pdbm.starter.business.messages.service;
 
+import de.pdbm.starter.business.messages.entity.Customer;
 import de.pdbm.starter.business.messages.entity.Product;
 import jakarta.ejb.Stateless;
+import jakarta.faces.application.FacesMessage;
+import jakarta.faces.context.FacesContext;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import jakarta.persistence.TypedQuery;
 
 import java.io.Serializable;
+import java.util.ArrayList;
 import java.util.List;
 
 import static jakarta.persistence.PersistenceContextType.TRANSACTION;
@@ -50,7 +54,38 @@ public class ProductService implements Serializable {
     }
 
     public void delete(Product product) {
-        em.remove(product);
+        if (!em.contains(product)) {
+            product = em.merge(product);
+        }
+        List<Long> referencedProductIds = this.getReferencedProductId(product);
+        if (referencedProductIds != null && !referencedProductIds.isEmpty()){
+
+            FacesContext.getCurrentInstance().addMessage(null, new FacesMessage(FacesMessage.SEVERITY_WARN,
+                    "Fehler", ": das Customer ist referenced by diese Orders, Sie können das nicht einfach wegmachen.Bitte setzen Sie die Customer_id von diesen Orders" + referencedProductIds + "auf null bevor Sie es löschen: " ));
+        } else {
+            em.remove(product);
+            FacesContext.getCurrentInstance().addMessage(null, new FacesMessage(FacesMessage.SEVERITY_INFO,
+                    "Erfolg", "Kunde erfolgreich gelöscht."));
+        }
+    }
+    public List<Long> getReferencedProductId(Product product){
+        TypedQuery<Long> query1 = em.createQuery(
+                "select s.stockPk FROM Stock s where s.stockPk.product_id = :product",Long.class
+        );
+        query1.setParameter("product", product.getId());
+        List<Long> result1 = query1.getResultList();
+        TypedQuery<Long> query2 = em.createQuery(
+                "select o.orderItemPk FROM OrderItem o where o.product = :product",Long.class
+        );
+
+        query2.setParameter("product", product);
+        List<Long> result2 = query2.getResultList();
+
+        // Zusammenführen der Ergebnisse
+        List<Long> combinedResults = new ArrayList<>(result1);
+        combinedResults.addAll(result2);
+
+        return combinedResults;
     }
 
     public void deleteById(Integer id) {
